Add --rescue and --cleanup
Add command line arguments to rescue an installed Stratis root
system and to clean up the chroot mounts.

When --rescue is given stratify will mount the root file system
and all its child mounts and start a bash shell in the chroot.
On exit from the shell the chroot is cleaned up and unmounted.