Try to re-use git builds for host and target system
Historically --git only built stratisd once: for the target system.
Since the addition of --git-build and --git-install the script will
build stratisd twice. Once for the host system and once for the target.
This is inefficient and time consuming: re-use the previous build and
install to the target system root.

This has two notable consequences:

  - When using --git there is no longer a checked-out git tree with
    a build present on the target system.

  - The host and target systems must be ABI compatible for the rust
    build. The easiest way to achieve this is to use the same
    distribution for both host and target.

The first could be fixed straightforwardly by copying the build tree
from the host to the target while finalizing the installation.